fix bugs

+++ app/src/main/java/com/mdground/hdenergy/models/ProjectWorkFuel.java
@@ -11,12 +11,12 @@ import java.util.List;
  */
 public class ProjectWorkFuel implements Parcelable{
 
-    private int AdjustInventory;    // 燃料调整
+    private float AdjustInventory;    // 燃料调整
     private String CreatedTime;
-    private int CurrentInventory;   // 本期库存
+    private float CurrentInventory;   // 本期库存
     private int FuelID;     // 燃料标识
     private String FuelName;// 燃料名称
-    private int PreviousInventory;  // 上期库存
+    private float PreviousInventory;  // 上期库存
     private List<ProjectFuelWarehouse> ProjectFuelWarehouseList; // 进料量列表
     private int WorkFuelID;
     private int WorkFurnaceID;
@@ -27,12 +27,12 @@ public class ProjectWorkFuel implements Parcelable{
     }
 
     protected ProjectWorkFuel(Parcel in) {
-        AdjustInventory = in.readInt();
+        AdjustInventory = in.readFloat();
         CreatedTime = in.readString();
-        CurrentInventory = in.readInt();
+        CurrentInventory = in.readFloat();
         FuelID = in.readInt();
         FuelName = in.readString();
-        PreviousInventory = in.readInt();
+        PreviousInventory = in.readFloat();
         ProjectFuelWarehouseList = in.createTypedArrayList(ProjectFuelWarehouse.CREATOR);
         WorkFuelID = in.readInt();
         WorkFurnaceID = in.readInt();
@@ -58,11 +58,11 @@ public class ProjectWorkFuel implements Parcelable{
     public void setAdjustReason(String adjustReason) {
         AdjustReason = adjustReason;
     }
-    public int getAdjustInventory() {
+    public float getAdjustInventory() {
         return AdjustInventory;
     }
 
-    public void setAdjustInventory(int adjustInventory) {
+    public void setAdjustInventory(float adjustInventory) {
         AdjustInventory = adjustInventory;
     }
 
@@ -74,11 +74,11 @@ public class ProjectWorkFuel implements Parcelable{
         CreatedTime = createdTime;
     }
 
-    public int getCurrentInventory() {
+    public float getCurrentInventory() {
         return CurrentInventory;
     }
 
-    public void setCurrentInventory(int currentInventory) {
+    public void setCurrentInventory(float currentInventory) {
         CurrentInventory = currentInventory;
     }
 
@@ -98,11 +98,11 @@ public class ProjectWorkFuel implements Parcelable{
         FuelName = fuelName;
     }
 
-    public int getPreviousInventory() {
+    public float getPreviousInventory() {
         return PreviousInventory;
     }
 
-    public void setPreviousInventory(int previousInventory) {
+    public void setPreviousInventory(float previousInventory) {
         PreviousInventory = previousInventory;
     }
 
@@ -137,12 +137,12 @@ public class ProjectWorkFuel implements Parcelable{
 
     @Override
     public void writeToParcel(Parcel parcel, int i) {
-        parcel.writeInt(AdjustInventory);
+        parcel.writeFloat(AdjustInventory);
         parcel.writeString(CreatedTime);
-        parcel.writeInt(CurrentInventory);
+        parcel.writeFloat(CurrentInventory);
         parcel.writeInt(FuelID);
         parcel.writeString(FuelName);
-        parcel.writeInt(PreviousInventory);
+        parcel.writeFloat(PreviousInventory);
         parcel.writeTypedList(ProjectFuelWarehouseList);
         parcel.writeInt(WorkFuelID);
         parcel.writeInt(WorkFurnaceID);
